# public/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "JoinGroups Communities API", "version": "1.0.0" },
  "servers": [{ "url": "https://joingroups.lat" }],
  "paths": {
    "/api/communities": {
      "get": {
        "operationId": "listCommunities",
        "summary": "Lista comunidades por tipo",
        "parameters": [
          {
            "in": "query",
            "name": "tipo",
            "required": true,
            "schema": { "type": "string", "enum": ["telegram","whatsapp","discord","facebook","clash-royale"] },
            "description": "Tipo de comunidad"
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": { "type": "string", "enum": ["nuevos","top","recientes"] },
            "description": "Orden de resultados"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": { "type": "integer", "default": 6, "minimum": 1, "maximum": 50 },
            "description": "Máximo de resultados"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["name","url","category"],
                    "properties": {
                      "name": { "type": "string" },
                      "url": { "type": "string", "format": "uri" },
                      "category": { "type": "string" },
                      "language": { "type": "string" },
                      "members": { "type": "integer" },
                      "description": { "type": "string" },
                      "createdAt": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clash/clanes": {
      "get": {
        "operationId": "listClashRoyaleClans",
        "summary": "Lista todos los clanes de Clash Royale almacenados",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 },
            "description": "Máximo de resultados a devolver"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "acceptTerms": { "type": "boolean" },
                          "link": { "type": "string", "format": "uri" },
                          "city": { "type": "string" },
                          "destacado": { "type": "boolean" },
                          "slug": { "type": "string" },
                          "categories": { "type": "string" },
                          "name": { "type": "string" },
                          "tag": { "type": "string" },
                          "tipo": { "type": "string" },
                          "description": { "type": "object" },
                          "createdAt": { "type": ["string", "null"] }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}